stress-zlib: print args.stream_bytes as unsigned long long
Use %lld and cats to unsigned long long, cleans up build warning on 32
bit platforms.

+++ stress-zlib.c
@@ -1277,9 +1277,9 @@ static int stress_zlib_deflate(
 	if (zlib_args.window_bits > 31)
 		zlib_args.window_bits = zlib_args.window_bits - 32;
 
-	pr_dbg("DEF: lvl=%d mem-lvl=%d wbits=%d strategy=%d stream-bytes=%ld\n",
+	pr_dbg("DEF: lvl=%d mem-lvl=%d wbits=%d strategy=%d stream-bytes=%llu\n",
 		zlib_args.level, zlib_args.mem_level, zlib_args.window_bits,
-		zlib_args.strategy, zlib_args.stream_bytes);
+		zlib_args.strategy, (unsigned long long)zlib_args.stream_bytes);
 	do {
 		if (stream_def.zalloc != Z_NULL) {
 			(void)deflateEnd(&stream_def);
